Hi-Res Logos!
Note: the file labeled PRIMARY is the one we should be using in most instances. There's a clone of that one on a white background which can also be used whenever you don't want a transparent background. The others are more 'just in case', but I'm thinking we won't end up needing them.